requirements.txt
litellm
python-dotenv
pytest
pytest-mock
pytest-cov
pytest-asyncio